cube: Turn on gpu validation via extension
Add code to use the VK_EXT_validation_flags to turn on
GPU validation.  It is disabled now because it does cause
a validation error because I was using an enum outside of
the range defined in the specification, as a temporary hack.
For experimentation, before we have support in the spec for
this, it is easier to use the layer settings file since there
is a lot more flexibility there.